pyproject.toml
[project]
name = "tdeck-notepad-tools"
version = "0.1.0"
description = "Host-side automation utilities for T-Deck firmware development."
requires-python = ">=3.10"
dependencies = [
    "opencv-python>=4.10.0.84",
    "pyserial>=3.5",
]

[tool.uv]
package = false
